spotify-streams-realtime-pipeline/requirements.txt
# Spotify Streams Real-Time Pipeline Python Dependencies

# PySpark & Data Processing
pyspark==3.5.0
pyarrow==11.0.0

# AWS SDKs
boto3==1.26.137
botocore==1.29.137

# Configuration & Environment
python-dotenv==1.0.0

# Testing
pytest==7.3.1
pytest-cov==4.0.0

# Code Quality
black==23.3.0
ruff==0.0.250
mypy==1.0.0
